Day 2022/16 part 2: some more speedups
If remaining is <= 2, then moving somewhere is not going to change
anything: if with remaining == 2 you do move, you will only have
remaining of i, then you will have to turn, and once the valve is
open, there will be 0 remaining, so the valve will release 0
pressure.

With this change, on the given example, we go from 2m33s to 1m49s.